backend/requirements.txt
amqp==5.3.1
anyascii==0.3.3
asgiref==3.9.2
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.3.1
django-filter==23.5
django-modelcluster==6.4
django-otp==1.6.1
django-permissionedforms==0.1
django-phonenumber-field==8.1.0
django-stubs-ext==5.2.5
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
kombu==5.5.4
laces==0.1.2
MarkupSafe==3.0.3
msgpack==1.1.1
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
pillow_heif==1.1.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
redis==6.4.0
requests==2.32.5
sendgrid==6.12.5
setuptools==80.1.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wagtail==7.1.1
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.6.0
Willow==1.11.0
